finish module
