+++ packages/@n8n/backend-common/src/modules/__tests__/module-registry.test.ts
@@ -34,6 +34,7 @@ describe('eligibleModules', () => {
 			'sso-saml',
 			'log-streaming',
 			'ldap',
+			'quick-connect',
 		]);
 	});
 
@@ -54,6 +55,7 @@ describe('eligibleModules', () => {
 			'sso-saml',
 			'log-streaming',
 			'ldap',
+			'quick-connect',
 		]);
 	});
