+++ packages/@n8n/backend-common/src/modules/module-registry.ts
@@ -44,6 +44,7 @@ export class ModuleRegistry {
 		'sso-saml',
 		'log-streaming',
 		'ldap',
+		'quick-connect',
 	];
 
 	private readonly activeModules: string[] = [];
