+++ packages/@n8n/backend-common/src/modules/modules.config.ts
@@ -17,6 +17,7 @@ export const MODULE_NAMES = [
 	'sso-saml',
 	'log-streaming',
 	'ldap',
+	'quick-connect',
 ] as const;
 
 export type ModuleName = (typeof MODULE_NAMES)[number];
