+++ packages/cli/src/modules/quick-connect/quick-connect.module.ts
@@ -1,12 +1,17 @@
 import type { ModuleInterface } from '@n8n/decorators';
-import { BackendModule, OnShutdown } from '@n8n/decorators';
+import { BackendModule } from '@n8n/decorators';
+import { Container } from '@n8n/di';
 
 @BackendModule({ name: 'quick-connect' })
 export class QuickConnectModule implements ModuleInterface {
-	async init() {
-		await import('./quick-connect.controller');
+	/**
+	 * Settings exposed to the frontend under `/rest/module-settings`.
+	 *
+	 * The response shape will be `{ options: [{ packageName: string, credentialType: string, text: string, quickConnectType: string }]}`.
+	 */
+	async settings() {
+		const { QuickConnectConfig } = await import('./quick-connect.config');
+		const { options } = Container.get(QuickConnectConfig);
+		return { options };
 	}
-
-	@OnShutdown()
-	async shutdown() {}
 }
