--- packages/cli/src/modules/quick-connect/quick-connect.service.ts
@@ -1,12 +1,0 @@
-import { Service } from '@n8n/di';
-
-import { QuickConnectConfig } from './quick-connect.config';
-
-@Service()
-export class QuickConnectService {
-	constructor(private readonly config: QuickConnectConfig) {}
-
-	getConfiguredQuickConnectOptions() {
-		return this.config.options;
-	}
-}
